Abstracted login logic to AuthHook.js
