+++ client/src/App.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useEffect } from "react";
+import React from "react";
 import {
   BrowserRouter as Router,
   Route,
@@ -12,39 +12,12 @@ import UserPlaces from "./places/pages/UserPlaces";
 import MainNavigation from "./shared/components/navigation/MainNavigation";
 import UpdatePlace from "./places/pages/UpdatePlace";
 import Auth from "./user/pages/Auth";
+import useAuth from './shared/components/hooks/AuthHook';
 import { AuthContext } from "./shared/components/context/AuthContext";
 
+
 const App = () => {
-  const [token, setToken] = useState(false);
-  const [loggedInUserId, setloggedInUserId] = useState();
-
-  const login = useCallback((uid, token, expirationDate) => {
-    setToken(token);
-    setloggedInUserId(uid);
-    const tokenExpirationDate = expirationDate || new Date(new Date().getTime() + 1000 * 50 * 60);
-
-    localStorage.setItem('userData', JSON.stringify({
-      userId: uid,
-      token: token,
-      expirationDate: tokenExpirationDate.toISOString()
-    }));
-  }, []);
-
-  const logout = useCallback(() => {
-    setToken(null);
-    setloggedInUserId(null);
-    localStorage.removeItem('userData');
-  }, []);
-
-  useEffect(() => {
-    const storedData = JSON.parse(localStorage.getItem('userData'));
-    if (storedData
-      && storedData.token
-      && new Date(storedData.expirationDate) > new Date()
-    ) {
-      login(storedData.userId, storedData.token, new Date(storedData.expirationDate));
-    }
-  }, [login]);
+  const {token, login, logout, loggedInUserId} = useAuth();
 
   let routes;
